+++ plugins/ucc_chat/lib/ucc_chat/services/side_nav_service.ex
@@ -26,7 +26,6 @@ defmodule UccChat.SideNavService do
   end
 
   def render_more_channels(user_id) do
-    # user = Helpers.get_user! user_id, preload: [:roles, user_roles: :role]
     user = Accounts.get_user! user_id, preload: [:roles, user_roles: :role]
 
     channels = ChannelService.get_side_nav_rooms(user)